Remove opencv from IMAGE_INSTALL in sr-image.
opencv isn't available as a package in the install.  Use opencv-apps
to pull it in.  This installs a couple of opencv utilities that we're
not interested in.  They don't take up too much space, so I'm not that
bothered by them at the moment.

I didn't find an existing way of convincing the shlib code that it
wants to install library packages that aren't used by another
package.  Maybe I'll add a method of doing that at some point...