Change OE tree to base based on release-2010.12 tree
The stable/2009 branch is now just too old.  Maintaining that is
comparable to the pain of rebasing on something more recent.  So, this
commit rebases on release-2010.12 (via the release-2010-12-11
branch in our openembedded.git repo).

In order to allow an incremental transition over to this new
situation, this is:

 a) Residing in a separate branch for a short while.
 b) Disabling the psplash stuff, as the way OE works with this has
    significantly changeed since stable/2009 was begun.  It'll be back
    later.