Add a package containing some ssh host keys
This is part of my recent campaign to allow the image that comes out
of OE to be booted read-only without having to be booted rw once.

Since dropbear's host keys are generated on first-boot, it cannot
write them to the read-only filesystem.  This patch adds a package,
called "sr-sshkey", which contains some already-generated keys.

This of course means that all the boards have the same host key.  This
is in no way a problem.  It has always been the case in the past
anyway!